# openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Router API
    version: 0.0.1
paths:
    /router/exactInputSingle:
        post:
            tags:
                - Router
            operationId: Router_ExactInputSingle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.router.v1.ExactInputSingleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.router.v1.ExactInputSingleReply'
    /router/exactOutputSingle:
        post:
            tags:
                - Router
            operationId: Router_ExactOutputSingle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.router.v1.ExactOutputSingleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.router.v1.ExactOutputSingleReply'
components:
    schemas:
        api.router.v1.ExactInputSingleReply:
            type: object
            properties:
                amountOut:
                    type: string
        api.router.v1.ExactInputSingleRequest:
            type: object
            properties:
                tokenIn:
                    type: string
                tokenOut:
                    type: string
                fee:
                    type: integer
                    format: uint32
                recipient:
                    type: string
                deadline:
                    type: string
                amountIn:
                    type: string
                amountOutMinimum:
                    type: string
                priceLimit:
                    type: string
        api.router.v1.ExactOutputSingleReply:
            type: object
            properties:
                amountIn:
                    type: string
        api.router.v1.ExactOutputSingleRequest:
            type: object
            properties:
                tokenIn:
                    type: string
                tokenOut:
                    type: string
                fee:
                    type: integer
                    format: uint32
                recipient:
                    type: string
                deadline:
                    type: string
                amountOut:
                    type: string
                amountInMaximum:
                    type: string
                priceLimit:
                    type: string
tags:
    - name: Router